pyproject.toml
[project]
name = "doc-workflow-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django[argon2]>=6.0.3",
    "djangorestframework>=3.17.0",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "drf-spectacular>=0.29.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
]
